I am struggling with the separation of the events board and the smart scheduler. Ideally the worflow would be Events are input into the databse through the events modal. As those events are bieing added a photographer is automatically added based on availability and number of hours already worked as most days should be 10 hours or less.
I believe the shot requests should have their own modal/page as there is much editing and arrangement that should occur. Lets try to segregate functions that are more event input and scheduling related and those that are shot list input and accountability focused.

In reality, much of this is related to the planning phase of an event. When an event is underway and we are actively taking in images and looking to fulfill deliverables there is a whole other set of functionality and dasjhboards that could be helpful. Is it worth taking the navugation back a step further and organize modals into plan, shoot, edit, deliver?