I want to make sure my prior request was addressed:
THis is the module that will most closely interact with the ingestion utility as it will be creating cards based on each ingestion done by the various editors. The data on these card should be derived from the reporting gathered and submitted by the ingestion utility at the local level.